Fix: Onboarding 스크린 및 뷰모델 통합
