+++ app/src/main/java/com/acon/acon/navigation/nested/OnboardingNavigation.kt
@@ -4,15 +4,21 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.navigation
+import com.acon.feature.onboarding.LinkedHashSetNavType
+import com.acon.feature.onboarding.OnboardingResultNavType
 import com.acon.feature.onboarding.OnboardingRoute
-import com.acon.feature.onboarding.screen.FrequentPlaceSelectScreen.composable.FrequentPlaceSelectScreenContainer
+import com.acon.feature.onboarding.StringListNavType
+import com.acon.feature.onboarding.StringSetNavType
+import com.acon.feature.onboarding.screen.OnboardingScreen.OnboardingResult
+import com.acon.feature.onboarding.screen.delete.FrequentPlaceSelectScreen.composable.FrequentPlaceSelectScreenContainer
 import com.acon.feature.onboarding.screen.OnboardingScreen.composable.OnboardingContainer
-import com.acon.feature.onboarding.screen.PerferredPlaceRateScreen.composable.PreferredPlaceRateScreenContainer
+import com.acon.feature.onboarding.screen.delete.PerferredPlaceRateScreen.composable.PreferredPlaceRateScreenContainer
 import com.acon.feature.onboarding.screen.PrefResultLoadingScreen.composable.PrefResultLoadingScreenContainer
-import com.acon.feature.onboarding.screen.PreferredFoodRateScreen.composable.PreferredFoodRateScreenContainer
-import com.acon.feature.onboarding.screen.PreferredPlaceSelectScreen.composable.PreferredPlaceSelectScreenContainer
-import com.acon.feature.onboarding.screen.UnlikeFoodSelectScreen.composable.UnlikeFoodScreenContainer
+import com.acon.feature.onboarding.screen.delete.PreferredFoodRateScreen.composable.PreferredFoodRateScreenContainer
+import com.acon.feature.onboarding.screen.delete.PreferredPlaceSelectScreen.composable.PreferredPlaceSelectScreenContainer
+import com.acon.feature.onboarding.screen.delete.UnlikeFoodSelectScreen.composable.UnlikeFoodScreenContainer
 import com.acon.feature.spot.com.acon.feature.spot.SpotRoute
+import kotlin.reflect.typeOf
 
 
 internal fun NavGraphBuilder.onboardingNavigationNavigation(
@@ -22,72 +28,14 @@ internal fun NavGraphBuilder.onboardingNavigationNavigation(
     navigation<OnboardingRoute.Graph>(
         startDestination = OnboardingRoute.OnboardingScreen
     ) {
-        composable<OnboardingRoute.ChooseUnlikeFoods> {
-            UnlikeFoodScreenContainer(
-                navigateToNextPage = {
-                    navController.navigate(OnboardingRoute.RatePreferFoods)
-                },
-                navigateToSpotListView = {
-                    navController.navigate(SpotRoute.SpotList)
-                }
+        composable<OnboardingRoute.LastLoading>(
+            typeMap = mapOf(
+                typeOf<Set<String>>() to StringSetNavType,
+                typeOf<List<String>>() to StringListNavType,
+                typeOf<LinkedHashSet<String>>() to LinkedHashSetNavType,
+                typeOf<OnboardingResult>() to OnboardingResultNavType,
             )
-        }
-
-        composable<OnboardingRoute.RatePreferFoods> {
-             PreferredFoodRateScreenContainer(
-                 navigateToPreviousPage = {
-                     navController.popBackStack()
-                 },
-                 navigateToNextPage = {
-                     navController.navigate(OnboardingRoute.SelectFrequentPlace)
-                 },
-                 navigateToSpotListView = {
-                     navController.navigate(SpotRoute.SpotList)
-                 }
-             )
-
-
-        }
-        composable<OnboardingRoute.SelectFrequentPlace> {
-            FrequentPlaceSelectScreenContainer(
-                navigateToPreviousPage = {
-                    navController.popBackStack()
-                },
-                navigateToNextPage = {
-                    navController.navigate(OnboardingRoute.SelectPreferPlace)
-                },
-                navigateToSpotListView = {
-                    navController.navigate(SpotRoute.SpotList)
-                }
-            )
-        }
-        composable<OnboardingRoute.SelectPreferPlace> {
-            PreferredPlaceSelectScreenContainer(
-                navigateToPreviousPage = {
-                    navController.popBackStack()
-                },
-                navigateToNextPage = {
-                    navController.navigate(OnboardingRoute.RatePreferPlace)
-                },
-                navigateToSpotListView = {
-                    navController.navigate(SpotRoute.SpotList)
-                }
-            )
-        }
-        composable<OnboardingRoute.RatePreferPlace> {
-            PreferredPlaceRateScreenContainer(
-                 navigateToPreviousPage = {
-                     navController.popBackStack()
-                 },
-                 navigateToNextPage = {
-                     navController.navigate(OnboardingRoute.LastLoading)
-                 },
-                navigateToSpotListView = {
-                    navController.navigate(SpotRoute.SpotList)
-                }
-             )
-        }
-        composable<OnboardingRoute.LastLoading> {
+        ) {
             PrefResultLoadingScreenContainer(
                 navigateToSpotListView = {
                     navController.navigate(SpotRoute.SpotList)
@@ -97,7 +45,7 @@ internal fun NavGraphBuilder.onboardingNavigationNavigation(
         composable<OnboardingRoute.OnboardingScreen> {
             OnboardingContainer(
                 navigateToLoadingView = {
-                    navController.navigate(OnboardingRoute.LastLoading)
+                    navController.navigate(OnboardingRoute.LastLoading(it))
                 },
                 navigateToSpotListView = {
                     navController.navigate(SpotRoute.SpotList)
