+++ feature/onboarding/src/main/java/com/acon/feature/onboarding/OnboardingRoute.kt
@@ -1,30 +1,17 @@
 package com.acon.feature.onboarding
 
+import com.acon.feature.onboarding.screen.OnboardingScreen.OnboardingResult
 import kotlinx.serialization.Serializable
 
-@Serializable
 sealed interface OnboardingRoute {
 
     @Serializable
     data object Graph : OnboardingRoute
 
     @Serializable
-    data object ChooseUnlikeFoods : OnboardingRoute
-
-    @Serializable
-    data object RatePreferFoods : OnboardingRoute
-
-    @Serializable
-    data object SelectFrequentPlace : OnboardingRoute
-
-    @Serializable
-    data object SelectPreferPlace : OnboardingRoute
-
-    @Serializable
-    data object RatePreferPlace : OnboardingRoute
-
-    @Serializable
-    data object LastLoading : OnboardingRoute
+    data class LastLoading(
+        val onboardingResult: OnboardingResult
+        ) : OnboardingRoute
 
     @Serializable
     data object OnboardingScreen : OnboardingRoute
