--- feature/onboarding/src/main/java/com/acon/feature/onboarding/screen/PreferredPlaceSelectScreen/composable/PreferredPlaceSelectScreenContainer.kt
@@ -1,2 +1,0 @@
-package com.acon.feature.onboarding.screen.PreferredPlaceSelectScreen.composable
-
