+++ feature/onboarding/src/main/java/com/acon/feature/onboarding/screen/delete/UnlikeFoodSelectScreen/UnlikeFoodSelectScreenViewModel.kt
@@ -1,4 +1,4 @@
-package com.acon.feature.onboarding.screen.UnlikeFoodSelectScreen
+package com.acon.feature.onboarding.screen.delete.UnlikeFoodSelectScreen
 
 import com.acon.core.utils.feature.base.BaseContainerHost
 import com.acon.domain.repository.OnboardingRepository
